+++ eric_sse/entities.py
@@ -10,7 +10,7 @@ from eric_sse.exception import InvalidListenerException, NoMessagesException
 logger = eric_sse.get_logger()
 
 MESSAGE_TYPE_CLOSED = '_eric_channel_closed'
-MESSAGE_TYPE_END_OF_STREAM = '_eric_channel_closed'
+MESSAGE_TYPE_END_OF_STREAM = '_eric_channel_eof'
 MESSAGE_TYPE_INTERNAL_ERROR = '_eric_error'
 
 
@@ -90,7 +90,7 @@ class AbstractChannel(ABC):
 
     def __init__(self, stream_delay_seconds: int = 0):
         """
-        :param stream_delay_seconds: Can be used to limit response rate of streamings. Only applies to message_stream calls.
+        :param stream_delay_seconds: Can be used to limit response rate of streaming. Only applies to message_stream calls.
         """
         logger.debug(f'Creating channel {AbstractChannel.NEXT_ID}')
         with Lock():